+++ FootballManager/FootballManager.Domain/Entities/Player.cs
@@ -9,10 +9,10 @@ namespace FootballManager.Domain.Entities
     {
         public string FirstName { get; set; }
         public string Surname { get; set; }
-        public int IdNo { get; set; }
+        public Int64 IdNo { get; set; }
         public DateTime DateOfBirth { get; set; }
-        public int Height { get; set; }
-        public int Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Weight { get; set; }
         public TeamPlayerAssignment TeamAssignment { get; set; }
 
     }
